refactor: extract embedding batch helpers
